Delete package-lock.json
Security problems with socket.io-parser library: "Due to improper type validation in the socket.io-parser library (which is used by the socket.io and socket.io-client packages to encode and decode Socket.IO packets), it is possible to overwrite the _placeholder object which allows an attacker to place references to functions at arbitrary places in the resulting query object."

Will look for another way of using a parser instead of this package that always had security issues. Better remove it than have security problems inside the program.